Completed redesign of wire spool holders.
Code should be cleaned up.  And if you choose certain combinations
of parameters, it won't work because of a bug in rounded_polygon
or because there's too much hardcoding with the DIN mount design.

Still, it's very usable with 3mm wall thickness and at least 22mm
spool widths.  Slices fine at 0.2 and 0.3 layer heights.