[AIDGEAR] Added a backing parameter to gear definitions.
The backing is the height-to-pitch-line for a rack or the radial
pitch-circle-to-outer-circle for a ring gear.  For a spur gear,
it's unnecessary and thus set to `undef` (not 0) in order for
the default to be set appropriately when a spur gear is the
template for a rack or ring gear definition.