[AIDGEAR] Improved handling of the "backing" parameter.
The backing defines the height (to the pitch line) of a rack or the
radial distance from the pitch line to the outer diameter for an
internal (ring) gear.  In order to provide appropriate defaults,
some calculation had to be moved out of the universal gear definition
function and into the type-specific ones.

The backing of the internal gears is now part of the polygon itself,
so there's less special-case code.