[AIDDIN] Checking in the current mess.
I'm still looking for a good way to secure the mount to the rail
that's low profile.  I've tried several iterations of a sliding
locking tab, but none of those print reliably because of the tight
spacing between the moving and non-moving parts.

The wire spool holders used vertical snap tabs, which work well, but
they use quite a bit of vertical space, which I'm trying to avoid
here.

Next up, I might revisit ideas for a pivoting latch.