Improved reliability of spider dropper.
Reduced diameter of spacer so that it doesn't interfere with the teeth
on the driver gear.

Switched from herringbone to helical teeth so the transition from no
engagement to full-engagement isn't so sensitive to slight offsets.

Raised guide for string to reduce risk of string coming off the spool
and getting caught up in the teeth.