Increased the depth of the spool and adjusted the guide height.
The deeper groove in the spool better handles the thickness of
several turns of string (such as the elastic string I'm trying).
I removed the recently added bias in the guide height, so it
should be aligned with the middle of the spool instead of the
top edge.

Also, after testing with the M3x16 flanged screws, I decided it
was okay to recess the flanged heads to the same depth as the
regular heads, simplifying the mounting holes for the deer motor.

This commit also contains some NFC refactorings, like the slots
on the bracket.