Changed gear ratio from 55:11 to 55:12.
Read a reminder that using relatively prime numbers of teeth in a
gear chain will help distribute wear evenly.  GCD(55, 12) == 1.

Also, because of the risk of undercuts, having too few teeth on a
smaller gear is bad.  Bumping 11 to 12 isn't a big change, but it's
in the rigth direction.

The result is that the spool is a bit larger.