[FASTRAK] Improve the locking tabs and lower the text.
The features of the lever snaps on the sides of the tab were too small
to print well.  I also extended the block in the middle to help align
the tabs with the slots before stress is applied ot the lever snaps.

The text is key to providing friction that holds the pass in the
bracket, but it was so high that it created stress on the folds.