Introduced aidpinball.scad.
This may not be general purpose enough to justify living among my
OpenSCAD libraries long term.  I created it to support some of Rex's
pinball machine design ideas, somewhat as a proof of concept.  We are
interested in having a single source for designing both the parts
mounted to the playfield and the playfield board (with appropriate
cutouts).