[AIDGEAR] Improved handling of "mutilated" gears.
A gear definition now includes the set of teeth that should be
"depopulated."  By default, that's an empty set, and you get a
regular gear.  You can specify the depopulation set when defining
the gear, or you may make a multilated (or un-multilated) version
of an existing gear definition.

Spur gears work as before, though a few more points are generated
to keep the gear rounder where there are missing teeth.

Ring gears work as before, though parts with missing teeth are now
on the base circle rather than the root circle.

Racks can now be mutilated using the same interface.

The spider dropper was updated to work with the more general API.
A couple unrelated tweaks to the dropper design were included in
this commit.