Improved the spool assembly.
Made the flange more printer-friendly, with 45-degree overhangs.

Eliminated the post in the pocket because the post always snapped off.
Also eliminated the pocket because it was too small to grab the end of
the string.  Instead, I made a channel that routes the string to the
surface of the spool, and a button for tying off the end.