[SPIDERDROPPER] Started adding idler arm, but improved guide instead.
I think there's enough promise in the idler arm to do additional
experiments, but I want a stable version for demoing in a couple days.

The guide is now stronger, and the sharp edges into the hole have been
rounded off because they were causing the kite string to fray.