[SPIDERDROPPER] Experimented with a better tension arm.
I think the tension arm is probably a dead end, but I've left the code
for it commented out for now.

The arm performs almost as well as the fixed guide.  To my surprise,
weighing it down did not limit its rotation, so it would immediate
start interferring with the drive gear, and it did nothing to dampen
the bounce when the spider drops.