[FASTRAK] Incrementally improved Fastrak e-tag holder.
The length was short by 1 mm, causing stress that turned to deformation
when heated.

Re-oriented upright for printing, which I hypothesize may help with how
it copes with heat build-up inside the structure.

I added a logo window so you can see that it's a FasTrak e-tag from
outside the car.  This also reduces the total amount of materials and
increases the surface area to volume ratio for better heat dissipation.