[AIDTHREAD] Improved fidelity to the standard.
* External threads are now true to size (no nozzle_d hack).
* External threads no longer overextend past the major diameter.
* Internal threads do extend to the major diameter.

Having the internal threads cut a little deeper gives good results with
a tighter tolerance, so there's less jiggle.

There may have been a bug fix for being 2-manifold.

Note that rendering sometimes gives a "PolySet has nonplanar faces"
warning.  That's not a problem with the generated polyhedron.  It
happens when intersecting with the cube to cut the threads cylinder to
the desired height (length).